CROSSTOOLS: wrapped_clang: smarter argument wrapping
Sometimes wrapped_clang will get called with a single argument
like:

  '-Xlinker -add_ast_path -Xlinker a/path'

and my previous simplistic "split on the first space" was failing.

Now the splitting tries to be smarter and peel out the flags and
the arguments by looking at whether the first character is a dash
but this is still fragile and will break on a path with spaces
and dashes.

I don't see how to fix it here though... I think Bazel should
do proper argument quoting at a much higher layer than here.